Directly sync XMPP client presence with WhatsApp
This commit removes period background synchronization for own presence,
and insteads synchronizes XMPP client presence with WhatsApp directly;
this in turn allows our state on WhatsApp to be more indicative of what
the actual use state is (assuming no use of official WhatsApp clients).

Synchronization of user status messages is also enabled, but only if
these are set to a non-empty value -- reverting to an empty status
message requires intervention via an official WhatsApp client.