Fix background presence sync timer
Previous iterations of the presence sync timer would `nil` out timers
when our presence isn't set to "available" when they fire, so that we
can pick up after them once we go to that status once again. However,
this was subtly broken as we were reading from the timer in between
those events, which led to (non-fatal) nil dereference errors.