Use raw bytes for Go-Python data interchange
This commit moves back to using raw, in-memory byte representations for
data interchange between Go and Python, where possible; previously,
conversions between Go `[]byte` and Python `bytes` types (and vice-versa)
were too slow, as they were handled as slices on an index-by-index basis.

Nevertheless, we still use temporary files for audio/video conversion,
as FFmpeg has trouble converting to/from certain formats (like MP4)
when streaming, as it needs to seek to different parts of the data
stream. This situation may be improved in the future, as we either move
to a different implementation, or solve issues in a different way, e.g.
memory-mapped files.