Add support for location messages
This commit adds basic support for location messages, which on the
WhatsApp side come with rich metadata (which is, as-of-yet, not exposed
to XMPP), while on the XMPP side are represented as simple `geo` URIs.

As noted above, incoming messages are mostly stripped of metadata down
to their latitude, longitude, and accuracy, as most XMPP clients will
refuse to parse messages containing anything other than a `geo` URI as
location messages. Future commits might have us expose metadata for
incoming location messages as separate XMPP messages, but outgoing
messages will remain quite plain for the time being.