Refactor and simplify media conversion
Media conversion is handled via FFmpeg (and FFprobe for metadata)
collection, which require a hefty set of command-line arguments for any
specific format we're looking to target. This has, until this point,
been handled in an attachment-centric manner, with little separation of
concerns between different layers of semantics.

This commit splits media handling (including conversion, metadata
extraction, thumbnailing, and waveform calculation) into a separate
package, which itself abstracts away FFmpeg internals, and allows us to
specify a target "specification"; thus, rather than assuming that any
audio file is being converted to Opus-encoded Ogg, we now allow for the
explicit specification of output format and parameters.

This helps make the various integrations more robust (and potentially
testable), and will allow for converting incoming media to
better-supported formats in the future.