Implement outgoing Youtube inline previews
Links to YouTube videos seem to be supported as inline previews on
WhatsApp clients; this commit implements marking any message containing
a link to a YouTube video as being of type "video", which has WhatsApp
display these with inline previews on official clients. More domains may
be added in the future, as support emerges.

In addition to the above, `media.Convert()` calls given images can now
specify image dimensions, which will have images be resized as they're
converted.

Closes: #40