Implement partial support for Broadcasts
Broadcasts is a WhatsApp feature for sending mass messages to many
(up to 256) contacts, without including them in a group-chat (which
might otherwise expose all their phone numbers amongst one another).

These messages appear inline with other 1-to-1 discussions by the
sending user, and are only distinguished by a small megaphone icon
accompanying the message.

This commit implements simple support for incoming broadcasts, with the
caveat of not supporting broadcast carbons or outgoing broadcasts (which
don't have a perfect analogue on the XMPP side).

Future support might have us fix bugs and expand with an ad-hoc command
for outgoing broadcasts, but simple support should suffice in not losing
messages.