media: Consolidate thumbnail generation
Thumbnail generation is now handled under the generic `Convert`
function and use of the `ImageWidth` and `ImageHeight` fields of the
`Spec` struct, with `GetThumbnail` being removed entirely.

In addition to this, we no longer assume, but rather explicitly check
the MIME type of incoming data, ensuring that conversions requested are
indeed valid (e.g. no converting an audio file to an image).